+++ pkg/cmd/pulumi/refresh.go
@@ -45,6 +45,7 @@ func newRefreshCmd() *cobra.Command {
 	var skipPreview bool
 	var suppressOutputs bool
 	var suppressPermaLink bool
+	var printSelfManagedBackendPermalink bool
 	var yes bool
 	var targets *[]string
 
@@ -89,6 +90,16 @@ func newRefreshCmd() *cobra.Command {
 				Type:                 displayType,
 				EventLogPath:         eventLogPath,
 				Debug:                debug,
+			}
+
+			filestateBackend, err := isFilestateBackend(opts.Display)
+			if err != nil {
+				return result.FromError(err)
+			}
+			// by default, we are going to suppress the permalink when using non-SaaS backends
+			// this can be re-enabled using the `print-self-managed-backend-permalink` flag
+			if filestateBackend && !printSelfManagedBackendPermalink {
+				opts.Display.SuppressPermaLink = true
 			}
 
 			s, err := requireStack(stack, true, opts.Display, true /*setCurrent*/)
@@ -196,6 +207,9 @@ func newRefreshCmd() *cobra.Command {
 	cmd.PersistentFlags().BoolVar(
 		&suppressPermaLink, "suppress-permalink", false,
 		"Suppress display of the state permalink")
+	cmd.PersistentFlags().BoolVar(
+		&printSelfManagedBackendPermalink, "print-self-managed-backend-permalink", false,
+		"Print self managed backend permalinks e.g. S3, Azure blob, GCS Storage etc.")
 	cmd.PersistentFlags().BoolVarP(
 		&yes, "yes", "y", false,
 		"Automatically approve and perform the refresh after previewing it")